chenged name of class, added import StringUtils, test code is formatted
